@@ -23,8 +23,8 @@ import { GeneratedTransaction } from '../../../../services/wallet-operations/tra
   styleUrls: ['./send-preview.component.scss'],
 })
 export class SendVerifyComponent implements OnDestroy {
-  @ViewChild('sendButton', { static: false }) sendButton: ButtonComponent;
-  @ViewChild('backButton', { static: false }) backButton: ButtonComponent;
+  @ViewChild('sendButton') sendButton: ButtonComponent;
+  @ViewChild('backButton') backButton: ButtonComponent;
   // Transaction which is going to be shown.
   @Input() transaction: GeneratedTransaction;
   // Emits when the preview must be removed from the UI and the form must be shown again. The
