@@ -22,7 +22,7 @@ export class ChangeNoteComponent implements OnInit, OnDestroy {
   // Max chars the note can have.
   public static readonly MAX_NOTE_CHARS = 64;
 
-  @ViewChild('button', { static: false }) button: ButtonComponent;
+  @ViewChild('button') button: ButtonComponent;
   form: FormGroup;
   maxNoteChars = ChangeNoteComponent.MAX_NOTE_CHARS;
   // Deactivates the form while the system is busy.
